drivers/spi: Allowing user to specify DCX pins from device tree.
Pin is specified in the respective device tree node for SPI, And
read by the nRF SPIM driver during build time. If the option is
specified, it is passed to the nRF library. If it is not available,
we pass 0xFF to tell the library that this support is not required
for the particular instance.

The user has the capability to specify the "dcx-pin" property for
any nRF SPI node, but it will only be read for instances for which
"NRFX_SPIM_EXTENDED_ENABLED" is defined (only true for SPI3 on
nRF52840)

Signed-off-by: Bilal Wasim <bilalwasim676@gmail.com>