subsys/fs: Adding notion of "dynamic" flash map.
This commit adds a couple of APIs which allows the user to create
their own custom partition table, store that table in the flash.

A sample application is added which tries to create partitions and
read them.

The partition table is currently stored in the "image-1" flash partition
and needs to be moved to its own partition.

We also need to add code in flash map APIs (get_flash_area_from_id, flash_area_foreach)
to read a partition from flash instead of using the hardcoded table.

Still some work to do here..

Signed-off-by: Bilal Wasim <bilalwasim676@gmail.com>