subsys/flash: Updating dynamic partition APIs to use Settings backend
Instead of managing the pages itself, update the dynamic partition APIs
to use the Settings backend. The APIs expect that the settings interface
has already been set up and so they don't do any set up.

Instead of registering callbacks with the subsystem and handling
loading/storing of data from NVS ourselves, we are trying to utilize as
much internal infrastructure as possible. Therefore, all writes to the
flash are done using the "settings_save_one" call and all values are
loaded "immediately".

In terms of keys, we have:
  .. partitions/cnt: Used to store the number of partitions
  .. partitions/${n}: Used to store info about a particular partition
                      At max, we can store info about 10 partitions.

In terms of user APIs, we expect the user to always add a new partition
by using the "add_partition_at_end" interface. If the user wants to modify/
overwrite existing partitions, they can do so using add_partition_by_id or
add_partition_at_index function, where the former probably makes more sense.

To comply with the new scheme, the Flash Map API(s) have been updated to
find the partition in flash, and only look at the DT table if the partition
is not found.

Signed-off-by: Bilal Wasim <bilalwasim676@gmail.com>