drivers/sensor: Adding more infrastructure for the BNO08x driver
This commit adds the device tree binding, and basic init code for
the BNO08x driver and a simple application to test all of this.

The application just inits SH2 and tries to read the product IDs.
More logic will be added in the app and the HAL as we make progress.
The app logic has been copied from the Hillcrest example app with
STM32: https://github.com/hcrest/bno080-nucleo-demo

Though we have most of the infrastructure to support multiple instance
in place, the HAL APIs exposed by the Hillcrest library doesn't
expose parameters to support multi-instance support, and I imagine
this is not important anyway.

Signed-off-by: Bilal Wasim <bilalwasim676@gmail.com>