drivers/spi: Base infrastructure for Hardware SPI CSN
Adding basic infrastructure to support Hardware Chip Select in the nRF
SPIM driver. Like DCX, this also introduces a new DT binding, which is
used to specify the pin to be used by the SPI module. If the user doesn't
want to use hardware chip select, they should not provide the "csn-pin"
in their SPI device tree node, and use the alternative "cs-gpios" binding.

Signed-off-by: Bilal Wasim <bilalwasim676@gmail.com>