drivers/spi: Hack to get SPI+BNO08x working
The BNO driver does a single transaction in two parts:
  .. Firstly, it sends 2 bytes and gets 2 bytes back in response. These
  response bytes tell the BNO driver the total length of the data to be
  recvd from the BNO chip.
  .. Using the length extracted from the first step, a new SPI transfer
  is created to tx/rx the remaning data.
The Zephyr SPI driver treats these transactions as seperate, and
therefore asserts/deasserts CS for each transfer individually. This
causes the BNO chip to misbehave.

To fix this problem, a very bad hack has been added to the driver which
ensures that CSN is asserted only after the 2nd transfer. This means
that this driver is basically unusable for other devices. Moreover, we
are unable to use the hardware csn with this approach as well.

Signed-off-by: root <root@root.com>