drivers/sensor: Finalise and clean up the BNO driver HAL
This commit updates the BNO driver to work properly with Zephyr SPI.
Most of the driver implementation has been copied from the bno nucleo
demo, and is adapted to use zephyr apis instead.

Tested with a demo app and works. Testing is kept minimal at this point.

Some minor parts of the driver need some improvement (signified by
TODOs) in the code.

Signed-off-by: root <root@root.com>